Improved ownership and thread responsibility.
- Improved ownership

Something I didn't like was the fact that I hold an Arc<Mutex<_>> for
the handlers list, just because the HttpServer had the methods to add
handlers and also had the thread logic to call the requests.

Because nothing stopped anyone from manipulating the handlers at the
same time the HttpServer was running, Rust compiler forced me to have
this Arc<Mutex<_>. The complete type was something like:
Arc<Mutex<HashMap<(String, String), Arc<RequestHandler>>>>

The inner handler was also under the Arc, because the thread pool
accessed the map and also copied the handler to launch the request,
while dropped the hash map to allow multiple request in parallel.

That smelled nasty. So by providing an HttpBuilder I can then
manipulate the HashMap while no thread is running, and when calling
build(), consume the HttpBuilder, move it's handlers to the HttpServer,
and the HttpServer would not need to hold the Arc<Mutex<_>> over the
hash map. Now the new map signature is just:
HashMap<(String, String), Arc<RequestHandler>>

- Better thread responsibility

Another thing that I didn't like was that the TcpStream was on hold
while the request was beign parsed and mapped to a handler. I don't know
if it's overkill, but I've decided that the parsing and dispatching of a
request should go to another thread while the main thread just receives
TCP streams and delegates it's parsing and routing to another thread
doing the dispatcher work.

With this, the HashMap from the HttpServer get's moved to the
HttpDispatcher, the HttpServer then just loops over TcpStreams that it
receives while .incoming(), and sends those streams to another thread
through the HttpDispatcher.

Is then the dispatcher that owns the ThreadPool, parses the request,
matches the handler, and spawns the handler call in the ThreadPool.

- Other oddities added

The body can now be read from the HttpRequest lazily. To achieve this, the
HttpRequest now owns the TcpStream to still be able to get the body
while it's beign processed.

This had a flaw. The response cannot be written because the TcpStream is
moved inside the HttpRequest. To allow retrieving the TcpStream again,
HttpRequest implements a trait to .unpack_stream(), which consumes the
HttpRequest and returns ownership of the TcpStream.

With this change, we can get the TcpStream, but the HttpRequest was lost
due to beign moved inside the handler.

To avoid this, we also need to change our handlers closure signature to
get a shared reference for the HttpRequest, instead of owning it.

Now our dispatcher still has ownership, and can call .unpack_stream()
once the handler has finished processing the HttpRequest, and then send
to the same TcpStream the data from the HttpResponse the handler
provided.