Removed ThreadPoolError and fixed error messages.
As with the previous commit, I was going to reimplement the
ThreadPoolError to a proper enum. The error was use once, so I've
decided to remove the custom error and just propagate the Sender error
directly.

Finally, I've got some error messages wrong. The .lock() error and
.recv() error were swapped, and also not handled properly.

Now they are handled properly, with the PosionedMutex error from
.lock() beign ignored now, as it means a previous request panicked
and in theory we could continue processing requests.